Add s3 bucket sqs (#41)
* Add s3 bucket sqs

* Add stripe fairing

* Remove unnecessary match

* Remove redundant prefixes

+++ src/fairings/sqs.rs
@@ -1,35 +1,23 @@
 use crate::aws;
 use crate::config::Config;
-use crate::guards::DbConn;
-use diesel::prelude::*;
 use rocket::fairing::{Fairing, Info, Kind};
 use rocket::{tokio, Orbit, Rocket};
 use rusoto_core::HttpClient;
 use rusoto_core::Region;
+mod s3_bucket;
+mod stripe;
+use self::stripe::StripeHandler;
+use crate::guards::DbConn;
 use rusoto_sqs::{DeleteMessageRequest, ReceiveMessageRequest, Sqs, SqsClient};
-use serde::Deserialize;
+use s3_bucket::S3BucketHandler;
 
 pub struct SqsFairing;
 
-#[derive(Deserialize)]
-struct RecordS3Object {
-  key: String,
-}
-
-#[derive(Deserialize)]
-struct RecordS3 {
-  object: RecordS3Object,
-}
-
-#[derive(Deserialize)]
-struct Record {
-  s3: RecordS3,
-}
-
-#[derive(Deserialize)]
-struct SqsMessageBody {
-  #[serde(rename = "Records")]
-  records: Vec<Record>,
+#[async_trait]
+pub trait QueueHandler: Send + Sync {
+  fn new(db_conn: DbConn, config: Config) -> Self;
+  fn url(&self) -> String;
+  async fn handle(&self, message: &rusoto_sqs::Message) -> ();
 }
 
 impl SqsFairing {
@@ -38,16 +26,22 @@ impl SqsFairing {
   }
 
   async fn init(&self, rocket: &Rocket<Orbit>) {
+    self.poll::<S3BucketHandler>(rocket).await;
+    self.poll::<StripeHandler>(rocket).await;
+  }
+
+  async fn poll<T: QueueHandler>(&self, rocket: &Rocket<Orbit>) {
     let db_conn = DbConn::get_one(&rocket)
       .await
       .expect("Failed to get db connection");
 
-    let config = rocket.state::<Config>().unwrap();
-    let queue_url = config.s3_bucket_queue.clone();
+    let config = rocket.state::<Config>().unwrap().clone();
     let aws_access_key_id = config.aws_access_key_id.clone();
     let aws_secret_key = config.aws_secret_key.clone();
 
     tokio::spawn(async move {
+      let handler = T::new(db_conn, config);
+
       let client = SqsClient::new_with(
         HttpClient::new().unwrap(),
         aws::CredentialsProvider::new(aws_access_key_id, aws_secret_key),
@@ -56,45 +50,24 @@ impl SqsFairing {
 
       loop {
         let receive_request = ReceiveMessageRequest {
-          queue_url: queue_url.clone(),
+          queue_url: handler.url(),
           wait_time_seconds: Some(20),
           ..Default::default()
         };
 
         let response = client.receive_message(receive_request).await.unwrap();
 
-        match response.messages {
-          Some(messages) => {
-            for message in messages {
-              use crate::schema::recordings;
-              use crate::schema::recordings::dsl::*;
+        if let Some(messages) = response.messages {
+          for message in messages {
+            handler.handle(&message).await;
 
-              let body = message.body.unwrap();
-              let message_body: SqsMessageBody = serde_json::from_str(&body).unwrap();
+            let delete_request = DeleteMessageRequest {
+              queue_url: handler.url(),
+              receipt_handle: message.receipt_handle.clone().unwrap(),
+            };
 
-              for record in message_body.records {
-                db_conn
-                  .run(move |conn| {
-                    let existing_recording =
-                      recordings::table.filter(recordings::video_key.eq(&record.s3.object.key));
-
-                    diesel::update(existing_recording)
-                      .set(uploaded.eq(true))
-                      .execute(conn)
-                      .unwrap();
-                  })
-                  .await;
-              }
-
-              let delete_request = DeleteMessageRequest {
-                queue_url: queue_url.clone(),
-                receipt_handle: message.receipt_handle.clone().unwrap(),
-              };
-
-              client.delete_message(delete_request).await.unwrap();
-            }
+            client.delete_message(delete_request).await.unwrap();
           }
-          None => {}
         }
       }
     });
